requirements.txt
aiogram==3.14.0
python-dotenv